+++ CHANGELOG.rst
@@ -15,6 +15,8 @@ Changes
 
 - Allow building with Cython 0.28 and above
 - Use `v` prefix on releases to have fixed links for this document
+- Add support for datetime
+- Add support for Decimal
 
 v0.4.8 - 2018-05-28
 ===================